requirements.txt
fastapi
uvicorn
pymongo
pandas
scikit-learn
numpy
python-dotenv
requests
